requirements.txt
# AI Pick Daily - Python Dependencies

# LLM APIs
google-generativeai>=0.8.0
anthropic>=0.40.0

# Data Sources
finnhub-python>=2.4.0

# Database
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0

# Technical Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
ta>=0.11.0

# Market Calendar
pandas-market-calendars>=4.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.0.0
ruff>=0.8.0
mypy>=1.0.0
